fix(): Fixed IP alloc after nse restart
After the vl3 nse restarts, it gets requests from existing NSCs with the
ip context filled out with the ip address that was previously assigned
to them. The vl3 nse should reserve those IPs before sending a response.
Without doing that, the IP pool is not update and hence any new pod coming up
will receive an IP which could be in use by other NSCs.

Signed-off-by: Bharath Horatti <bharath@aveshasystems.com>